Make window events with single handler
GLFW can only have a single callback for each type of event. If we want
to hook up multiple callbacks (maybe the user does), then we will need
to multiplex these event callbacks. Luckily, we can do so with
AddHandler. In the future, we can easily expose these handlers for users.