Simplify event handlers while making them general
We treat events similarly to drawing now, where the structure of the
event handling tree is not encoded in the types. This allows events to
be more general compared to the old system where we can only have a
single list of handlers, and they either handle an event or pass it on.