Set all pixel origins to bottom-left corner
GLFW uses the top-left corner to relay the mouse position. This caused a
bug in the clipping example because the origin used by glViewport is in
the bottom-left. I think that a top-left origin is silly anyway, so I
set the origin for mouse interaction to be in the bottom-left.